# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "CardCreate": {
        "additionalProperties": false,
        "properties": {
          "antonyms": {
            "default": [],
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 20,
            "title": "Antonyms",
            "type": "array"
          },
          "deck_id": {
            "format": "uuid",
            "title": "Deck Id",
            "type": "string"
          },
          "example_sentence": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Sentence"
          },
          "example_source": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Source"
          },
          "example_translation": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Translation"
          },
          "learned_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learned On"
          },
          "meaning": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Meaning",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "enum": [
                  "noun",
                  "verb",
                  "adjective",
                  "adverb",
                  "pronoun",
                  "determiner",
                  "preposition",
                  "conjunction",
                  "interjection",
                  "particle",
                  "auxiliary",
                  "numeral",
                  "phrase",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "part_of_speech_detail": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech Detail"
          },
          "pronunciation": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation"
          },
          "reading": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading"
          },
          "romanization": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Romanization"
          },
          "supplementary_note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplementary Note"
          },
          "synonyms": {
            "default": [],
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 20,
            "title": "Synonyms",
            "type": "array"
          },
          "target_language_definition": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Language Definition"
          },
          "term": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "term",
          "meaning",
          "deck_id"
        ],
        "title": "CardCreate",
        "type": "object"
      },
      "CardDetail": {
        "properties": {
          "antonyms": {
            "items": {
              "type": "string"
            },
            "title": "Antonyms",
            "type": "array"
          },
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deck": {
            "$ref": "#/components/schemas/DeckSummary"
          },
          "example_sentence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Sentence"
          },
          "example_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Source"
          },
          "example_translation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Translation"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "learned_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learned On"
          },
          "meaning": {
            "title": "Meaning",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "part_of_speech_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech Detail"
          },
          "pronunciation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation"
          },
          "reading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading"
          },
          "review_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewState"
              },
              {
                "type": "null"
              }
            ]
          },
          "romanization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Romanization"
          },
          "supplementary_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplementary Note"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "title": "Synonyms",
            "type": "array"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagSummary"
            },
            "title": "Tags",
            "type": "array"
          },
          "target_language_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Language Definition"
          },
          "term": {
            "title": "Term",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "deck",
          "term",
          "meaning",
          "reading",
          "pronunciation",
          "romanization",
          "part_of_speech",
          "archived_at",
          "version",
          "updated_at",
          "target_language_definition",
          "example_sentence",
          "example_translation",
          "example_source",
          "synonyms",
          "antonyms",
          "part_of_speech_detail",
          "note",
          "supplementary_note",
          "learned_on",
          "created_at",
          "tags",
          "review_state"
        ],
        "title": "CardDetail",
        "type": "object"
      },
      "CardSummary": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "deck": {
            "$ref": "#/components/schemas/DeckSummary"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "meaning": {
            "title": "Meaning",
            "type": "string"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "pronunciation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation"
          },
          "reading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading"
          },
          "romanization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Romanization"
          },
          "term": {
            "title": "Term",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "deck",
          "term",
          "meaning",
          "reading",
          "pronunciation",
          "romanization",
          "part_of_speech",
          "archived_at",
          "version",
          "updated_at"
        ],
        "title": "CardSummary",
        "type": "object"
      },
...
      "DeckSummary": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "explanation_language": {
            "enum": [
              "en",
              "ja",
              "zh-TW"
            ],
            "title": "Explanation Language",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "target_language": {
            "enum": [
              "en",
              "ja"
            ],
            "title": "Target Language",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "target_language",
          "explanation_language",
          "archived_at"
        ],
        "title": "DeckSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_CardSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[CardSummary]",
        "type": "object"
      },
...
      "ReviewState": {
        "properties": {
          "ease_factor": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Ease Factor",
            "type": "string"
          },
          "interval_days": {
            "title": "Interval Days",
            "type": "integer"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "next_review_at": {
            "format": "date-time",
            "title": "Next Review At",
            "type": "string"
          },
          "review_stage": {
            "title": "Review Stage",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "review_stage",
          "ease_factor",
          "interval_days",
          "last_reviewed_at",
          "next_review_at",
          "version"
        ],
        "title": "ReviewState",
        "type": "object"
      },
...
      "SemanticSearchItem": {
        "properties": {
          "deck_id": {
            "format": "uuid",
            "title": "Deck Id",
            "type": "string"
          },
          "distance": {
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Distance",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "meaning": {
            "title": "Meaning",
            "type": "string"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "pronunciation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation"
          },
          "reading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading"
          },
          "romanization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Romanization"
          },
          "score": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Score",
            "type": "number"
          },
          "term": {
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "id",
          "deck_id",
          "term",
          "meaning",
          "reading",
          "pronunciation",
          "romanization",
          "part_of_speech",
          "distance",
          "score"
        ],
        "title": "SemanticSearchItem",
        "type": "object"
      },
      "SemanticSearchRequest": {
        "additionalProperties": false,
        "properties": {
          "deck_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deck Id"
          },
          "limit": {
            "default": 10,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "maxLength": 500,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          },
          "target_language": {
            "const": "en",
            "title": "Target Language",
            "type": "string"
          }
        },
        "required": [
          "query",
          "target_language"
        ],
        "title": "SemanticSearchRequest",
        "type": "object"
      },
      "SemanticSearchResponse": {
        "properties": {
          "eligible_count": {
            "minimum": 0.0,
            "title": "Eligible Count",
            "type": "integer"
          },
          "index_status": {
            "enum": [
              "complete",
              "partial",
              "empty"
            ],
            "title": "Index Status",
            "type": "string"
          },
          "indexed_count": {
            "minimum": 0.0,
            "title": "Indexed Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SemanticSearchItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "index_status",
          "eligible_count",
          "indexed_count"
        ],
        "title": "SemanticSearchResponse",
        "type": "object"
      },
      "TagSummary": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "TagSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "English Learning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cards": {
      "get": {
        "description": "List compact owned card summaries with language and tag filtering.",
        "operationId": "list_cards_v1_cards_get",
        "parameters": [
          {
            "in": "query",
            "name": "deck_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deck Id"
            }
          },
          {
            "in": "query",
            "name": "target_language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "en",
                    "ja"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Language"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "archived",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CardSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Cards"
      },
      "post": {
        "operationId": "create_card_v1_cards_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Card"
      }
    },
    "/v1/cards/semantic-search": {
      "post": {
        "description": "Rank only current embeddings belonging to the authenticated owner.",
        "operationId": "semantic_search_v1_cards_semantic_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Semantic Search"
      }
    },
...
  }
}